+++ src/core/numeric.F90
@@ -29,7 +29,172 @@ module numeric
     real(realwp), parameter :: &
         pi = f4*atan(f1)
 
+    ! "Sensible" thresholds for scientific calculations
+    ! These thresholds are used internally for closeness tests.
+    ! near0: to test numbers in terms of numeric precision (epsilon)
+    ! small: for final quantities with values typically of magnitude around 1
+    ! null0: values below this thresholds can be safely considered null.
+    real(real32), parameter, private :: &
+        near0_r32 = 1.0e-6_real32, &
+        small_r32 = 1.0e-4_real32, &
+        null0_r32 = 1.0e-32_real32
+    real(real64), parameter, private :: &
+        near0_r64 = 1.0e-10_real64, &
+        small_r64 = 1.0e-6_real64, &
+        null0_r64 = 1.0e-32_real64
+        
+    interface operator(.iscloseto.)
+        module procedure :: is_close_to_r32_generic, is_close_to_r64_generic
+    end interface
+
+    interface is_close_to
+        module procedure :: is_close_to_r32, is_close_to_r64
+    end interface
+
 contains
+
+! ======================================================================
+
+elemental function is_close_to_r32(value, target, rel_tol, abs_tol) result(res)
+    !! Check if value is close to target within chosen tolerance(s).
+    !!
+    !! Checks if a given value is close to a target value within numeric
+    !! tolerances.  The tolerances can be defined as absolute or relative.
+    !! The test is the following:
+    !!
+    !! |value-target| <= max(rel_tol * max(|value|, |target|), abs_tol)
+    !!
+    !! @note
+    !! Negative numbers in tolerances values are ignored.
+    !! The default values are used in this case.
+    !! @endnote
+    real(real32), intent(in) :: value
+    !! Value to check.
+    real(real32), intent(in) :: target
+    !! Target for the closeness check.
+    real(real32), intent(in), optional :: rel_tol
+    !! Relative tolerance threshold, by default as epsilon(value)*10
+    real(real32), intent(in), optional :: abs_tol
+    !! Absolute tolerance threshold, by default as 10^-32.
+    !! The default value is chosen sensible in scientific applications.
+    logical :: res
+    !! result of the closeness test.
+
+    real(real32) :: tol_abs, tol_rel
+
+    if (present(rel_tol)) then
+        if (rel_tol < 0.0_real32) then
+            tol_rel = near0_r32
+        else
+            tol_rel = rel_tol
+        end if
+    else
+        tol_rel = near0_r32
+    end if
+    if (present(abs_tol)) then
+        if (abs_tol < 0.0_real32) then
+            tol_abs = null0_r32
+        end if
+        tol_abs = abs_tol
+    else
+        tol_abs = null0_r32
+    end if
+
+    res = abs(value-target) <= max(tol_rel * max(abs(value), abs(target)), &
+                                   tol_abs)
+end function is_close_to_r32
+
+! ======================================================================
+
+elemental function is_close_to_r32_generic(value, target) result(res)
+    !! Check if value is close to target within chosen tolerance(s).
+    !!
+    !! Checks if a given value is close to a target value within numeric
+    !! standard tolerances for scientific applications.
+    !!
+    !! |value-target| <= max(rel_tol * max(|value|, |target|), abs_tol)
+    real(real32), intent(in) :: value
+    !! Value to check.
+    real(real32), intent(in) :: target
+    !! Target for the closeness check.
+    logical :: res
+    !! result of the closeness test.
+
+    res = abs(value-target) <= max(near0_r32 * max(abs(value), abs(target)), &
+                                   null0_r32)
+end function is_close_to_r32_generic
+
+! ======================================================================
+
+elemental function is_close_to_r64(value, target, rel_tol, abs_tol) result(res)
+    !! Check if value is close to target within chosen tolerance(s).
+    !!
+    !! Checks if a given value is close to a target value within numeric
+    !! tolerances.  The tolerances can be defined as absolute or relative.
+    !! The test is the following:
+    !!
+    !! |value-target| <= max(rel_tol * max(|value|, |target|), abs_tol)
+    !!
+    !! @note
+    !! Negative numbers in tolerances values are ignored.
+    !! The default values are used in this case.
+    !! @endnote
+    real(real64), intent(in) :: value
+    !! Value to check.
+    real(real64), intent(in) :: target
+    !! Target for the closeness check.
+    real(real64), intent(in), optional :: rel_tol
+    !! Relative tolerance threshold, by default as epsilon(value)*10
+    real(real64), intent(in), optional :: abs_tol
+    !! Absolute tolerance threshold, by default as 10^-32.
+    !! The default value is chosen sensible in scientific applications.
+    logical :: res
+    !! result of the closeness test.
+
+    real(real64) :: tol_abs, tol_rel
+
+    if (present(rel_tol)) then
+        if (rel_tol < 0.0_real64) then
+            tol_rel = near0_r64
+        else
+            tol_rel = rel_tol
+        end if
+    else
+        tol_rel = near0_r64
+    end if
+    if (present(abs_tol)) then
+        if (abs_tol < 0.0_real64) then
+            tol_abs = null0_r64
+        else
+            tol_abs = abs_tol
+        end if
+    else
+        tol_abs = null0_r64
+    end if
+
+    res = abs(value-target) <= max(tol_rel * max(abs(value), abs(target)), &
+                                   tol_abs)
+end function is_close_to_r64
+
+! ======================================================================
+
+elemental function is_close_to_r64_generic(value, target) result(res)
+    !! Check if value is close to target within chosen tolerance(s).
+    !!
+    !! Checks if a given value is close to a target value within numeric
+    !! standard tolerances for scientific applications.
+    !!
+    !! |value-target| <= max(rel_tol * max(|value|, |target|), abs_tol)
+    real(real64), intent(in) :: value
+    !! Value to check.
+    real(real64), intent(in) :: target
+    !! Target for the closeness check.
+    logical :: res
+    !! result of the closeness test.
+
+    res = abs(value-target) <= max(near0_r64 * max(abs(value), abs(target)), &
+                                   null0_r64)
+end function is_close_to_r64_generic
 
 ! ======================================================================
 
@@ -54,28 +219,6 @@ function is_integer(arg) result(res)
     end select
 
 end function is_integer
-
-! ======================================================================
-
-function is_real(arg) result(res)
-    !! Check that argument is a real number.
-    !!
-    !! Checks that argument arg is a valid real number.
-    class(*), intent(in) :: arg
-    !! Argument to check.
-    logical :: res
-    !! Result of check.
-
-    select type (arg)
-        type is (real(real32))
-            res = .true.
-        type is (real(real64))
-            res = .true.
-        class default
-            res = .false.
-    end select
-
-end function is_real
 
 ! ======================================================================
 
@@ -104,6 +247,28 @@ function is_number(arg) result(res)
     end select
 
 end function is_number
+
+! ======================================================================
+
+function is_real(arg) result(res)
+    !! Check that argument is a real number.
+    !!
+    !! Checks that argument arg is a valid real number.
+    class(*), intent(in) :: arg
+    !! Argument to check.
+    logical :: res
+    !! Result of check.
+
+    select type (arg)
+        type is (real(real32))
+            res = .true.
+        type is (real(real64))
+            res = .true.
+        class default
+            res = .false.
+    end select
+
+end function is_real
 
 ! ======================================================================